this is an Auth provider, so I renamed useAuth to AuthProvider.
